+++ packages/env/CHANGELOG.md
@@ -11,6 +11,8 @@ between the host, web container, and cli container. If you still encounter permi
 
 -   Create an `afterSetup` option in `.wp-env.json` files for setting arbitrary commands to run after setting up WordPress when using `wp-env start` and `wp-env clean`.
 -   Add a `WP_ENV_AFTER_SETUP` environment variable to override the `afterSetup` option.
+-   Execute the `afterSetup` command on `wp-env start` after the environment is set up. This can happen when your config changes, WordPress updates, or you pass the `--update` flag.
+-   Execute the `afterSetup` command on `wp-env clean`.
 
 ### Bug fix